requirements.txt
fastapi[standard]
psycopg
psycopg-binary
sqlmodel
python-dotenv
pydantic-settings